Refactoring

+++ Connector/src/Main.ts
@@ -86,11 +86,13 @@ export async function main(process: NodeJS.Process): Promise<void> {
       await runConnector(bridgeJobDef, serverArgs, false, false);
     }
   } catch (error) {
-    process.stdout.write(error.message + "\n" + error.stack);
+    console.log(error.message + "\n" + error.stack);
   } finally {
     await Utilities.shutdownBackend();
-    IModelJsFs.purgeDirSync(KnownTestLocations.outputDir);
-    IModelJsFs.unlinkSync(path.join(KnownTestLocations.assetsDir, "test.db"));
+    if(IModelJsFs.existsSync(KnownTestLocations.outputDir))
+      IModelJsFs.purgeDirSync(KnownTestLocations.outputDir);
+    if(IModelJsFs.existsSync(path.join(KnownTestLocations.assetsDir, "test.db")))
+      IModelJsFs.unlinkSync(path.join(KnownTestLocations.assetsDir, "test.db"));
     process.exit();
   }
 }
@@ -113,10 +115,10 @@ const runConnector = async (bridgeJobDef: BridgeJobDefArgs, serverArgs: ServerAr
   imodel = await BriefcaseDb.open(new ClientRequestContext(), briefcases[0].key, { openAsReadOnly: true });
   // ConnectorTestUtils.verifyIModel(imodel, bridgeJobDef, isUpdate, isSchemaUpdate);
   briefcaseEntry!.openMode = OpenMode.ReadWrite;
-  console.log(`Executing query: SELECT devicetype FROM cbd.Device`);
+  console.log(`Executing query: SELECT devicetype FROM cbd.Device from Main`);
   for await (const row of imodel.query(`SELECT devicetype FROM cbd.Device`)) {
     console.log(row);
-  }
+ }
   imodel.close();
 };
